Refactor: PaymentRepository 포트 추상화(헥사고날 아키텍처)
