+++ src/main/java/com/jerry/ticketing/payment/application/PaymentQueryService.java
@@ -10,7 +10,7 @@ import com.jerry.ticketing.member.application.dto.domain.MemberDto;
 import com.jerry.ticketing.payment.application.dto.web.CreatePaymentDto;
 import com.jerry.ticketing.payment.application.dto.domain.PaymentDto;
 import com.jerry.ticketing.payment.domain.Payment;
-import com.jerry.ticketing.payment.infrastructure.repository.PaymentRepository;
+import com.jerry.ticketing.payment.domain.port.PaymentRepository;
 import com.jerry.ticketing.reservation.application.ReservationQueryService;
 import com.jerry.ticketing.reservation.application.dto.domain.ReservationDto;
 import lombok.RequiredArgsConstructor;
