+++ src/main/java/com/jerry/ticketing/payment/infrastructure/repository/PaymentJpaRepository.java
@@ -1,0 +1,10 @@
+package com.jerry.ticketing.payment.infrastructure.repository;
+
+import com.jerry.ticketing.payment.domain.Payment;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface PaymentJpaRepository extends JpaRepository<Payment, Long> {
+    Optional<Payment> findByOrderId(String orderId);
+}
