--- src/main/java/com/jerry/ticketing/payment/infrastructure/repository/PaymentRepository.java
@@ -1,19 +1,0 @@
-package com.jerry.ticketing.payment.infrastructure.repository;
-
-import com.jerry.ticketing.payment.domain.Payment;
-import jakarta.persistence.LockModeType;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Lock;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import java.util.Optional;
-
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
-
-    @Lock(LockModeType.PESSIMISTIC_WRITE)
-    @Query("SELECT p FROM Payment p WHERE p.orderId = :orderId")
-    Optional<Payment> findByOrderIdWithLock(@Param(("orderId")) String orderId);
-
-    Optional<Payment> findByOrderId(String orderId);
-}
