+++ src/main/java/com/jerry/ticketing/payment/infrastructure/repository/PaymentRepositoryAdapter.java
@@ -1,0 +1,30 @@
+package com.jerry.ticketing.payment.infrastructure.repository;
+
+import com.jerry.ticketing.payment.domain.Payment;
+import com.jerry.ticketing.payment.domain.port.PaymentRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+
+import java.util.Optional;
+
+@Repository
+@RequiredArgsConstructor
+public class PaymentRepositoryAdapter implements PaymentRepository {
+
+    private final PaymentJpaRepository jpaRepository;
+
+    @Override
+    public Payment save(Payment payment) {
+        return jpaRepository.save(payment);
+    }
+
+    @Override
+    public Optional<Payment> findById(Long id) {
+        return jpaRepository.findById(id);
+    }
+
+    @Override
+    public Optional<Payment> findByOrderId(String orderId) {
+        return jpaRepository.findByOrderId(orderId);
+    }
+}
